+++ devel/tinytest/test-subset-charclass.R
@@ -45,11 +45,11 @@ stri_subset_charclass(x, "[2]", negate = TRUE) <- "e"
 expect_identical(x, c("2", NA, "2"))
 
 x <- c("stringi R", "123", "ID456", "", NA)
-stri_subset_charclass(x, "[1]") <- c(NA, "8")
+suppressWarnings(stri_subset_charclass(x, "[1]") <- c(NA, "8"))
 expect_identical(x, c("stringi R", NA, "ID456", "", NA))
 
 x <- c(NA, "stringi R", "123", "", "ID456")
-stri_subset_charclass(x, "[1]", negate = TRUE) <- c("a", "b", "c", "d")
+suppressWarnings(stri_subset_charclass(x, "[1]", negate = TRUE) <- c("a", "b", "c", "d"))
 expect_identical(x, c(NA, "a", "123", "b", "c"))
 
 x <- c("stringi R", "123", "ID456", "", NA)
@@ -65,7 +65,7 @@ stri_subset_charclass(x, "[7]") <- c("a", "b")
 expect_identical(x, c("stringi R", "123", NA, "ID456", ""))
 
 x <- c("stringi R", NA, "173", "ID457", "7")
-stri_subset_charclass(x, "[7]") <- c("a", "b")
+suppressWarnings(stri_subset_charclass(x, "[7]") <- c("a", "b"))
 expect_identical(x, c("stringi R", NA, "a", "b", "a"))
 
 x <- c("stringi R", "173", "ID457", "7")